requirements.txt
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
cloudinary==1.40.0
Flask==3.0.0
Flask-WTF==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
python-dotenv==1.0.1
six==1.16.0
urllib3==2.2.1
Werkzeug==3.0.1
WTForms==3.1.2
Flask-SQLAlchemy==3.1.1 
Flask-Session==0.8.0
Flask-Login==0.6.3